Add 3 Linux Tools
As the name suggests, ipmitool and FreeIPMI are tools for IPMI, but we
can also get the power consumption of a bare metal machine through DCMI
(IPMI extension).

PowerTOP is mainly for laptops, but it cat estimate the power
consumption of each process based on the battery discharge information
obtained by ACPI and Linux perf events etc.

I think these tools are useful, but please reject this PR if they don't
fit the purpose of this list.