removing irrelevant companies
Happy to be challenged on this, but I don't see how either of these organisations are specifically related to green software? if we open the list up to all green organisations, then clearly it would need to be much larger (and I don't see the value in that - we should focus on our area of expertise).

Signed-off-by: Ben Logan <1737958+benlogan@users.noreply.github.com>